HolsteinServer/requirements.txt
diff-match-patch==20121119
Django==1.8.2
django-crispy-forms==1.4.0
django-datetime-widget==0.9.3
django-easy-pdf==0.1.0
django-filter==0.10.0
django-import-export==0.2.7
django-mathfilters==0.3.0
django-mssql==1.7
django-pure-pagination==0.2.1
django-wkhtmltopdf==2.0.2
djangorestframework==3.1.3
djangorestframework-bulk==0.2.1
html5lib==0.99999
jdcal==1.0
openpyxl==2.2.5
pdfkit==0.5.0
Pillow==2.8.2
PyPDF2==1.24
python-gcm==0.2
pytz==2015.4
pywin32==218
reportlab==3.2.0
requests==2.7.0
six==1.9.0
tablib==0.10.0
vboxapi==1.0
xhtml2pdf==0.0.6
xlrd==0.9.3
xlutils==1.7.1
xlwt==1.0.0
